backend/pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["mihail8531 <mihail8531@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.0"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
pydantic-settings = "^2.0.3"
pydantic = "^2.4.2"
aiofiles = "^23.2.1"
aioshutil = "^1.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
aioboto3 = "^12.1.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
pytest = "^7.4.3"
black = "^23.11.0"
pytest-asyncio = "^0.21.1"
types-aiofiles = "^23.2.0.0"
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
